Remove use of conform.StandardizeChartDirectory
conform.StandardizeChartDirectory appears to be unnecessary. I
did some testing and the packages/vendor/chartName/chart directory
looks the same when calling it as when not calling it for the f5/
nginx-ingress package, which has two overlay files.

From a semantic point of view, all it does is load and then write out
the *chart.Chart again. Assuming the code we use from the helm project
works as it should, it does nothing.

I think it was left in the codebase after it ceased to be necessary.
If it turns out to still be necessary we can add it back in the future.