Remove prepare subcommand
The intended use of the prepare subcommand was to pull an upstream
chart into the package directory so that patch files could be
created. However, patch functionality never worked in partner-charts-ci,
no packages/charts use patches in rancher/partner-charts, and there
are no plans to do so in the future.

Removing the prepare subcommand simplifies the interface for
the user of partner-charts-ci, and it allows for further simplification
of the partner-charts-ci codebase.