Remove stage argument from generateChanges
With the removal of the prepare subcommand, generateChanges is
always called with either its auto or stage argument true.
This means that the code inside a few if statements in generateChanges
always runs, which means that we can remove these if statements.

Another consequence is that we can remove the stage argument,
because we now always do the operations represented by the stage
subcommand.