Introduce ChartWrapper type to track whether charts have been modified
We cannot simply read a .tgz chart file and then write it out again,
because we lose the metadata for the files it contains. This means
that git views it as modified, even when the contents of the files
are exactly the same. Tracking whether a chart has been modified,
and only writing the .tgz file if it has been modified, solves this
problem.